reorg cpp files; move wal_replayer into wal
