+++ src/storage/disk_array.cpp
