+++ src/storage/disk_array_collection.cpp
