+++ src/storage/shadow_file.cpp
