--- src/storage/storage_structure/CMakeLists.txt
@@ -1,9 +1,0 @@
-add_library(kuzu_storage_structure
-        OBJECT
-        disk_array.cpp
-        disk_array_collection.cpp
-        overflow_file.cpp)
-
-set(ALL_OBJECT_FILES
-        ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:kuzu_storage_structure>
-        PARENT_SCOPE)
