+++ src/storage/table/list_column.cpp
