+++ src/storage/table/null_column.cpp
