+++ src/storage/table/string_chunk_data.cpp
