+++ src/storage/table/struct_chunk_data.cpp
