+++ src/storage/table/struct_column.cpp
