+++ src/storage/table/version_record_handler.cpp
