+++ src/storage/wal/CMakeLists.txt
@@ -1,8 +1,8 @@
 add_library(kuzu_storage_wal
         OBJECT
-        shadow_file.cpp
         wal.cpp
-        wal_record.cpp)
+        wal_record.cpp
+        wal_replayer.cpp)
 
 set(ALL_OBJECT_FILES
         ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:kuzu_storage_wal>
